avoid dependency on Carp
Carp’s output was different up to 1.24, so tests fail on old perls. I do
not wish to condone this upstream change by changing my tests, much less
condone it at the expense of users by depending on a newer Carp, which
leaves me to in-house the tiny fraction of functionality from Carp that
I actually need. Thanks to the Carp maintainer for his service.